+++ js/data.js
@@ -1,0 +1,1 @@
+// модуль, который создаёт данные;
